+++ docs/Changelog.md
@@ -39,6 +39,7 @@
   of backend code completion and dictionaries substitution
 * [#4339](https://github.com/TouK/nussknacker/pull/4339) Improvements: Don't fetch state for archived/unarchived scenario, return computed based on last state action
 * [#4343](https://github.com/TouK/nussknacker/pull/4343) Updated Flink 1.16.1 -> 1.16.2
+* [#4342](https://github.com/TouK/nussknacker/pull/4342) Improvement: Don't run custom action on archived scenario and fragment
 
 1.9.1 (24 Apr 2023)
 ------------------------